src/requirements.txt
boto3>=1.28.0
pandas>=2.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
